pyproject.toml
[project]
name = "mcpbot"
version = "0.1.0"
description = "An implementation of an MCP client and server, using FastAPI."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-cosmos>=4.9.0",
    "azure-identity>=1.21.0",
    "azure-keyvault-secrets>=4.9.0",
    "fastapi>=0.115.12",
    "jinja2>=3.1.6",
    "langchain-chroma>=0.2.3",
    "langchain-core>=0.3.56",
    "langchain-mcp-adapters>=0.0.9",
    "langchain-openai>=0.3.14",
    "langgraph>=0.3.34",
    "mcp>=1.6.0",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "pytz>=2025.2",
]
